Moved all functions to domUpates
